Update Pipeline_Versioning_with_W&B_Artifacts_(PyTorch).ipynb (#215)
* Update Pipeline_Versioning_with_W&B_Artifacts_(PyTorch).ipynb

Hi there,

If I'm correct, right now when computing the k-examples with the highest loss, the example uploads the images to wandb with an incorrect pred label because it's using the `predictions` tensor that is not ordered by which predictions had a bigger loss. The change I propose should solve this issue. 

It's also possible I'm missing something, so if I'm wrong I'd appreciate some feedback.

Have a good weekend!

* Fix hardest k examples method

* fix nbs

Co-authored-by: Daniel Vera Nieto <daniel.veranieto@inf.ethz.ch>
Co-authored-by: Thomas Capelle <tcapelle@pm.me>